rename: point util 함수 파일 이름 변경
